+++ example/lib/main.dart
@@ -455,10 +455,7 @@ class _MyAppState extends State<MyApp> with SingleTickerProviderStateMixin {
 
       final signedPreKeyPrivate = PrivateKey.generate();
       final signedPreKeyPublic = signedPreKeyPrivate.getPublicKey();
-      final bobPrivate = PrivateKey.deserialize(
-        bytes: bobIdentity.privateKey.toList(),
-      );
-      final signedPreKeySignature = bobPrivate.sign(
+      final signedPreKeySignature = bobIdentity.sign(
         message: signedPreKeyPublic.serialize().toList(),
       );
       final timestamp = BigInt.from(DateTime.now().millisecondsSinceEpoch);
@@ -476,7 +473,7 @@ class _MyAppState extends State<MyApp> with SingleTickerProviderStateMixin {
 
       final kyberKeyPair = KyberKeyPair.generate();
       final kyberPreKey = kyberKeyPair.getPublicKey();
-      final kyberPreKeySignature = bobPrivate.sign(
+      final kyberPreKeySignature = bobIdentity.sign(
         message: kyberPreKey.serialize().toList(),
       );
       final kyberPreKeyRecord = KyberPreKeyRecord.create(